refactor: rename report utils to error utils
