+++ src/main/java/com/example/accommodiq/services/ReviewServiceImpl.java
@@ -3,15 +3,13 @@ package com.example.accommodiq.services;
 import com.example.accommodiq.domain.Host;
 import com.example.accommodiq.domain.Review;
 import com.example.accommodiq.dtos.MessageDto;
-import com.example.accommodiq.dtos.ReviewDto;
-import com.example.accommodiq.dtos.ReviewRequestDto;
 import com.example.accommodiq.dtos.ReviewStatusDto;
 import com.example.accommodiq.enums.ReviewStatus;
 import com.example.accommodiq.repositories.ReviewRepository;
 import com.example.accommodiq.services.interfaces.IAccommodationService;
 import com.example.accommodiq.services.interfaces.IHostService;
 import com.example.accommodiq.services.interfaces.IReviewService;
-import com.example.accommodiq.utilities.ReportUtils;
+import com.example.accommodiq.utilities.ErrorUtils;
 import org.hibernate.exception.ConstraintViolationException;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -20,7 +18,7 @@ import java.util.Collection;
 import java.util.Date;
 import java.util.Optional;
 
-import static com.example.accommodiq.utilities.ReportUtils.throwNotFound;
+import static com.example.accommodiq.utilities.ErrorUtils.throwNotFound;
 
 @Service
 public class ReviewServiceImpl implements IReviewService {
@@ -79,7 +77,7 @@ public class ReviewServiceImpl implements IReviewService {
 //        return review;
 
         if (reviewId == 4L) {
-            ReportUtils.throwNotFound("reviewNotFound");
+            ErrorUtils.throwNotFound("reviewNotFound");
         }
 
         return new MessageDto("Review successfully deleted.");
@@ -99,7 +97,7 @@ public class ReviewServiceImpl implements IReviewService {
 //        allReviews.flush();
 
         if (reviewId == 4L) {
-            ReportUtils.throwNotFound("reviewNotFound");
+            ErrorUtils.throwNotFound("reviewNotFound");
         }
 
         return new Review(
