+++ src/main/java/com/example/accommodiq/utilities/ErrorUtils.java
@@ -6,7 +6,7 @@ import org.springframework.web.server.ResponseStatusException;
 
 import java.util.ResourceBundle;
 
-public class ReportUtils {
+public class ErrorUtils {
 
     private static final ResourceBundle bundle = ResourceBundle.getBundle("ValidationMessages", LocaleContextHolder.getLocale());
